+++ src/main/resources/db/migration/V01__pessoa.sql
@@ -1,5 +1,5 @@
 CREATE TABLE pessoa(
      id BIGINT PRIMARY KEY AUTO_INCREMENT,
      nome VARCHAR(100) NOT NULL,
-     indice_produtividade INTEGER NOT NULL
+     indice_produtividade DECIMAL(4,2) NOT NULL
 )
